Fixed timing issues.
Noticed a pattern that in presence of 0's, the output eventually becomes
all 0's or an alternating pattern of 1's and 0's.

All 0's were already handled, so, added logic to detect alternating
pattern, flip it if odd no. of hours are remaining and break the loop.

All test cases passing.